load goal information client side
